Organize encoding and decoding separately
